+++ test/integrationPackage/lambda-files.tests.js
@@ -3,7 +3,7 @@
 const path = require('path');
 const { expect } = require('chai');
 const fse = require('fs-extra');
-const { execSync } = require('../utils/child-process');
+const { execSync } = require('../utils/childProcess');
 const serverlessExec = require('../serverlessBinary');
 const { getTmpDirPath, listZipFiles } = require('../utils/fs');
